django-src/requirements.txt
astroid==1.6.3
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
defusedxml==0.5.0
Django==2.0.4
django-allauth==0.36.0
django-composite-foreignkey==1.0.1
django-cors-headers==2.2.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-markdown-deux==1.0.5
django-pagedown==1.0.4
django-rest-auth==0.9.3
django-rest-multiple-models==2.1.0
django-templated-mail==1.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djoser==1.1.5
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
markdown2==2.3.5
mccabe==0.6.1
oauthlib==2.0.7
Pillow==5.1.0
psycopg2==2.7.4
PyJWT==1.6.1
pylint==1.8.4
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
urllib3==1.22
wrapt==1.10.11